move reusable theme codes to theme and admin theme
