+++ src/recipe/Baked.Recipe.Service.Application/Theme/Admin/AdminThemeFeature.cs
@@ -1,14 +1,93 @@
 ﻿using Baked.Architecture;
 
+using static Baked.Theme.Admin.Components;
+using static Baked.Ui.Datas;
+
 namespace Baked.Theme.Admin;
 
-public class AdminThemeFeature : IFeature<ThemeConfigurator>
+public class AdminThemeFeature(IEnumerable<Page> _pages,
+    Action<ErrorPage>? _errorPageOptions = default,
+    Action<SideMenu>? _sideMenuOptions = default,
+    Action<Header>? _headerOptions = default
+) : IFeature<ThemeConfigurator>
 {
     public virtual void Configure(LayerConfigurator configurator)
     {
         configurator.ConfigureComponentExports(exports =>
         {
             exports.AddFromExtensions(typeof(Components));
+        });
+
+        configurator.ConfigureAppDescriptor(app =>
+        {
+            configurator.UsingLocalization(l =>
+            {
+                app.Error = ErrorPage(
+                    options: ep =>
+                    {
+                        ep.SafeLinks.AddRange([.. _pages.Where(p => p.ErrorSafeLink).Select(p => p.AsCardLink(l))]);
+                        ep.ErrorInfos[403] = ErrorPageInfo(l("Access Denied"), l("You do not have the permision to view the address or data specified."));
+                        ep.ErrorInfos[404] = ErrorPageInfo(l("Page Not Found"), l("The page you want to view is etiher deleted or outdated."));
+                        ep.ErrorInfos[500] = ErrorPageInfo(l("Unexpected Error"), l("Please contact system administrator."));
+
+                        _errorPageOptions.Apply(ep);
+                    },
+                    data: Computed(Composables.UseError)
+                );
+            });
+        });
+
+        configurator.ConfigureLayoutDescriptors(layouts =>
+        {
+            configurator.UsingLocalization(l =>
+            {
+                layouts.Add(DefaultLayout("default", options: dl =>
+                {
+                    dl.SideMenu = SideMenu(
+                        options: sm =>
+                        {
+                            sm.Menu.AddRange([.. _pages.Where(p => p.SideMenu).Select(p => p.AsSideMenuItem(l))]);
+
+                            _sideMenuOptions.Apply(sm);
+                        }
+                    );
+                    dl.Header = Header(options: h =>
+                    {
+                        foreach (var page in _pages)
+                        {
+                            if (page.Disabled) { continue; }
+
+                            h.Sitemap[page.Route] = page.AsHeaderItem(l);
+                        }
+
+                        _headerOptions.Apply(h);
+                    });
+                }));
+            });
+
+            layouts.Add(ModalLayout("modal"));
+        });
+
+        configurator.ConfigurePageDescriptors(pages =>
+        {
+            configurator.UsingDomainModel(domain =>
+            {
+                configurator.UsingLocalization(l =>
+                {
+                    foreach (var page in _pages)
+                    {
+                        if (page.Build is null) { continue; }
+
+                        pages.Add(page.Build(new()
+                        {
+                            Page = page,
+                            Sitemap = [.. _pages],
+                            Domain = domain,
+                            NewLocaleKey = l
+                        }));
+                    }
+                });
+            });
         });
     }
 }
