+++ src/recipe/Baked.Recipe.Service.Application/Theme/ThemeExtensions.cs
@@ -1,4 +1,6 @@
 ﻿using Baked.Architecture;
+using Baked.Domain.Model;
+using Baked.RestApi.Model;
 using Baked.Theme;
 
 namespace Baked;
@@ -13,5 +15,37 @@ public static class ThemeExtensions
         action?.Invoke(result);
 
         return result;
+    }
+
+    public static MethodModel GetMethod(this TypeModel type, string name) =>
+        type.GetMembers().Methods[name];
+
+    public static ActionModelAttribute GetAction(this MethodModel method) =>
+        method.GetSingle<ActionModelAttribute>();
+
+    public static string GetRoute(this ActionModelAttribute action, List<(string key, string value)> routeParameters)
+    {
+        var routeParts = action.GetRoute().Split("/");
+        foreach (var (key, value) in routeParameters)
+        {
+            var parameter = action.Parameter[key];
+            if (parameter.FromRoute)
+            {
+                routeParts[parameter.RoutePosition] = value;
+            }
+        }
+
+        return routeParts.Join('/');
+    }
+
+    public static IEnumerable<string> GetEnumNames(this TypeModel type) =>
+        [.. type.SkipNullable().Apply(t => Enum.GetNames(t).Select(n => n.TrimStart('_')))];
+
+    public static TypeModel SkipNullable(this TypeModel type)
+    {
+        if (!type.IsAssignableTo(typeof(Nullable<>))) { return type; }
+        if (!type.TryGetGenerics(out var generics)) { throw new InvalidOperationException($"{type.Name} doesn't provide generics information"); }
+
+        return generics.GenericTypeArguments.First().Model;
     }
 }
