+++ test/recipe/Baked.Test.Recipe.Service.Application/Locales/locale.tr.restext
@@ -1,13 +1,19 @@
 // package keys =//
 {0} with {1}: '{2}' does not exist=Bu {1}: '{2}' ile {0} yoktur
 '{0}' is not a valid {1}='{0}' geçerli bir {1} değildir
+Access Denied=Erişim Reddedildi
 An unexpected error has occured, please contact the administrator=Beklenmeyen bir hata oluştu lütfen sistem yöneticisiyle iletişime geçiniz
 An unexpected error has occured, please contact the administrator: '{0}'\n{1}=Beklenmeyen bir hata oluştu lütfen sistem yöneticisiyle iletişime geçiniz: '{0}'\n{1}
 Attempted to perform an unauthorized operation=Yetkisiz bir işlem gerçekleştirmeye çalışıldı
 Authentication=Kimlik Doğrulama
 Failed to authenticate with given credentials=Verilen kimlik bilgileriyle kimlik doğrulaması yapılamadı
+Page Not Found=Sayfa Bulunamadı
+Please contact system administrator.=Lütfen sistem yöneticinizle iletişime geçin.
 Record Not Found=Kayıt Bulunamadı
 Route Parameter Is Not Valid=Route Parametresi Geçerli Değil
+The page you want to view is etiher deleted or outdated.=Görüntülemek istediğiniz sayfa silinmiş veya güncel değil.
+Unexpected Error=Beklenmeyen Hata
+You do not have the permision to view the address or data specified.=Belirtilen adresi veya veriyi görüntüleme izniniz yok.
 
 // test project keys =//
 {0} should be unique={0} benzersiz olmalıdır
@@ -27,7 +33,6 @@ This should not exist in output locale files=Bunun locale çıktılarında bulun
 // turkish translations for frontend =//
 A component to display message=Mesaj göstermek için bir bileşen
 A component to give a link to a domain object=Bir alan nesnesine bağlantı vermek için bir bileşen
-Access Denied=Erişim Reddedildi
 Allow change site language=Site dilini değiştirmeye izin ver
 Allow custom configuration to define custom css and more=Özel CSS ve daha fazlasını tanımlamak için özelleştirmeye izin ver
 Allow locale customization and language support=Dil özelleştirmesi ve çoklu dil desteğine izin ver
@@ -97,10 +102,8 @@ Number=Sayı
 Optional=İsteğe Bağlı
 Page=Sayfa
 Page Title=Sayfa Başlığı
-Page Not Found=Sayfa Bulunamadı
 Parameter=Parametre
 Parameters=Parametreler
-Please contact system administrator.=Lütfen sistem yöneticinizle iletişime geçin.
 Plugins=Eklentiler
 Query Parameters=Sorgu Parametreleri
 Rate=Oran
@@ -130,9 +133,6 @@ Specs=Spesifikasyonlar
 String=Metin
 Sync and manage parameters in query string=Sorgu dizesindeki parametreleri senkronize et ve yönet
 The core component that renders a dynamic component using given descriptor=Verilen tanımlayıcıyı kullanarak dinamik bir bileşen oluşturan ana bileşen
-The page you want to view is etiher deleted or outdated.=Görüntülemek istediğiniz sayfa silinmiş veya güncel değil.
 Toast=Uyarı
 Try the links from the menu below to view the page you want to access=Erişmek istediğiniz sayfayı görmek için aşağıdaki menüden bağlantıları deneyin.
-Unexpected Error=Beklenmeyen Hata
 View list data in a table=Liste verilerini bir tabloda görüntüle
-You do not have the permision to view the address or data specified.=Belirtilen adresi veya veriyi görüntüleme izniniz yok.
