+++ test/recipe/Baked.Test.Recipe.Service.Application/Program.cs
@@ -1,5 +1,6 @@
 ﻿using Baked.Test.Orm;
 using Baked.Test.Theme.Custom;
+using Baked.Theme;
 
 Bake.New
     .Service(
