+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/CustomThemeExtensions.cs
@@ -1,5 +1,3 @@
-using Baked.Domain.Model;
-using Baked.RestApi.Model;
 using Baked.Test.Theme.Custom;
 using Baked.Theme;
 
@@ -10,36 +8,4 @@ public static class CustomThemeExtensions
     public static CustomThemeFeature Custom(this ThemeConfigurator _, Page indexPage,
         IEnumerable<Page>? pages = default
     ) => new([indexPage, .. pages ?? []]);
-
-    public static MethodModel GetMethod(this TypeModel type, string name) =>
-        type.GetMembers().Methods[name];
-
-    public static ActionModelAttribute GetAction(this MethodModel method) =>
-        method.GetSingle<ActionModelAttribute>();
-
-    public static string GetRoute(this ActionModelAttribute action, List<(string key, string value)> routeParameters)
-    {
-        var routeParts = action.GetRoute().Split("/");
-        foreach (var (key, value) in routeParameters)
-        {
-            var parameter = action.Parameter[key];
-            if (parameter.FromRoute)
-            {
-                routeParts[parameter.RoutePosition] = value;
-            }
-        }
-
-        return routeParts.Join('/');
-    }
-
-    public static IEnumerable<string> GetEnumNames(this TypeModel type) =>
-        [.. type.SkipNullable().Apply(t => Enum.GetNames(t).Select(n => n.TrimStart('_')))];
-
-    public static TypeModel SkipNullable(this TypeModel type)
-    {
-        if (!type.IsAssignableTo(typeof(Nullable<>))) { return type; }
-        if (!type.TryGetGenerics(out var generics)) { throw new InvalidOperationException($"{type.Name} doesn't provide generics information"); }
-
-        return generics.GenericTypeArguments.First().Model;
-    }
 }
