+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/DomainComponents.cs
@@ -1,5 +1,6 @@
 using Baked.Domain.Model;
 using Baked.RestApi.Model;
+using Baked.Theme;
 using Baked.Theme.Admin;
 using Baked.Ui;
 using Humanizer;
