+++ test/recipe/Baked.Test.Recipe.Service.Application/Theme/Custom/PageBuilders.cs
@@ -1,5 +1,6 @@
 ﻿using Baked.Business;
 using Baked.Test.Caching;
+using Baked.Theme;
 using Baked.Ui;
 
 using static Baked.Theme.Admin.Components;
